+++ CHANGES.rst
@@ -9,6 +9,11 @@
 - The oldest supported and tested Sphinx version is now 1.3.5. See
   :issue:`17`.
 
+- 100% test coverage.
+
+- Remove support for ``programoutput_use_ansi``. The
+  ``sphinxcontrib.ansi`` extension is no longer available on PyPI.
+
 0.8 (Oct 12, 2012)
 ==================